+++ packages/fields-auto-increment/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone/fields-auto-increment
+
 # @keystone-alpha/fields-auto-increment
 
 ## 2.0.5